[ONEM-24351] Split WebKitBrowser into the plugin and service libraries
Plugin split into two parts:
* libWPEFrameworkWebkitBrowser.so (framework part)
* libWPEFrameworkWebkitBrowserImpl.so (WPEProcess, service part)

Plugin json configuration files updated with expected locator variable
for WPEProcess.

WPEFramework::WebKitBrowser::MemoryObserverImpl moved.
