+++ WebKitBrowser/Amazon.config
@@ -4,6 +4,7 @@ set(preconditions Graphics Internet)
 
 map()
     kv(outofprocess true)
+    kv(locator lib${PLUGIN_WEBKITBROWSER_IMPLEMENTATION}.so)
     if(PLUGIN_AMAZON_USER)
         kv(user ${PLUGIN_AMAZON_USER})
     endif()
