+++ WebKitBrowser/Apps.config
@@ -4,6 +4,7 @@ set(preconditions Graphics Internet)
 
 map()
     kv(outofprocess true)
+    kv(locator lib${PLUGIN_WEBKITBROWSER_IMPLEMENTATION}.so)
 end()
 ans(rootobject)
