+++ WebKitBrowser/CMakeLists.txt
@@ -48,6 +48,8 @@ option(PLUGIN_WEBKITBROWSER_LIGHTNING_APP "Include Lightning App in seperate plu
 option(PLUGIN_WEBKITBROWSER_ENABLE_JIT "Enable the use of JIT javascript optimalization" ON)
 option(PLUGIN_WEBKITBROWSER_ENABLE_DFG "Enable the use of DFG javascript optimalization" ON)
 
+set(PLUGIN_WEBKITBROWSER_IMPLEMENTATION "${MODULE_NAME}Impl" CACHE STRING "Specify a library with a webkit implementation." )
+
 set(PLUGIN_WEBKITBROWSER_AUTOSTART false CACHE STRING "Automatically start WebKitBrowser plugin")
 set(PLUGIN_WEBKITBROWSER_TRANSPARENT false CACHE STRING "Set transparency")
 set(PLUGIN_WEBKITBROWSER_STARTURL "about:blank" CACHE STRING "Initial URL for WebKitBrowser plugin")
@@ -133,16 +135,7 @@ set(SOURCE_LIST
     Module.cpp
     WebKitBrowser.cpp
     WebKitBrowserJsonRpc.cpp
-    WebKitImplementation.cpp
 )
-
-if(NOT DEFINED WEBKIT_GLIB_API)
-    set(SOURCE_LIST
-        ${SOURCE_LIST}
-        InjectedBundle/Tags.cpp
-        InjectedBundle/Utils.cpp
-)
-endif()
 
 add_library(${MODULE_NAME} SHARED
     ${SOURCE_LIST}
@@ -156,16 +149,28 @@ target_link_libraries(${MODULE_NAME}
     PRIVATE
         CompileSettingsDebug::CompileSettingsDebug
         ${NAMESPACE}Plugins::${NAMESPACE}Plugins
-        ${NAMESPACE}Definitions::${NAMESPACE}Definitions
-        ${WPE_BACKEND_LIBRARIES}
-        ${LIBSOUP_LIBRARIES}
-        ${GLIB_LIBRARIES}
-        ${WPE_WEBKIT_LIBRARIES}
-        ${OPENSSL_LIBRARIES}
-        ${SECURITY_TOKEN_LIBRARY}
-        ${ODHERR})
+        ${NAMESPACE}Definitions::${NAMESPACE}Definitions)
 
 target_include_directories(${MODULE_NAME}
+    PRIVATE
+        ../helpers
+)
+
+add_library(${PLUGIN_WEBKITBROWSER_IMPLEMENTATION} SHARED
+    Module.cpp
+    WebKitImplementation.cpp)
+
+if(NOT DEFINED WEBKIT_GLIB_API)
+    target_sources(${PLUGIN_WEBKITBROWSER_IMPLEMENTATION} PRIVATE InjectedBundle/Utils.cpp)
+    target_sources(${PLUGIN_WEBKITBROWSER_IMPLEMENTATION} PRIVATE InjectedBundle/Tags.cpp)
+endif()
+
+set_target_properties(${PLUGIN_WEBKITBROWSER_IMPLEMENTATION} PROPERTIES
+        CXX_STANDARD 11
+        CXX_STANDARD_REQUIRED YES
+        FRAMEWORK FALSE)
+
+target_include_directories(${PLUGIN_WEBKITBROWSER_IMPLEMENTATION}
     PRIVATE
         ../helpers
         ${WPE_BACKEND_INCLUDE_DIRS}
@@ -176,16 +181,33 @@ target_include_directories(${MODULE_NAME}
         ${OPENSSL_INCLUDE_DIRS}
 )
 
+target_link_libraries(${PLUGIN_WEBKITBROWSER_IMPLEMENTATION}
+    PRIVATE
+        ${NAMESPACE}Plugins::${NAMESPACE}Plugins
+        ${NAMESPACE}Definitions::${NAMESPACE}Definitions
+        ${WPE_BACKEND_LIBRARIES}
+        ${WPE_WEBKIT_LIBRARIES}
+        ${OPENSSL_LIBRARIES}
+        ${ODHERR}
+)
+
 if(WPE_WEBKIT_DEPRECATED_API)
     target_compile_definitions(${MODULE_NAME} PRIVATE WPE_WEBKIT_DEPRECATED_API=${WPE_WEBKIT_DEPRECATED_API})
+    target_compile_definitions(${PLUGIN_WEBKITBROWSER_IMPLEMENTATION} PRIVATE WPE_WEBKIT_DEPRECATED_API=${WPE_WEBKIT_DEPRECATED_API})
 endif()
 
 if(WEBKIT_GLIB_API)
     target_compile_definitions(${MODULE_NAME} PRIVATE WEBKIT_GLIB_API=${WEBKIT_GLIB_API})
+    target_compile_definitions(${PLUGIN_WEBKITBROWSER_IMPLEMENTATION} PRIVATE WPE_WEBKIT_DEPRECATED_API=${WPE_WEBKIT_DEPRECATED_API})
 endif()
 
 install(
     TARGETS ${MODULE_NAME}
+    DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/${STORAGE_DIRECTORY}/plugins
+    COMPONENT ${MODULE_NAME})
+
+install(
+    TARGETS ${PLUGIN_WEBKITBROWSER_IMPLEMENTATION}
     DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/${STORAGE_DIRECTORY}/plugins
     COMPONENT ${MODULE_NAME})
