+++ WebKitBrowser/SearchAndDiscoveryApp.config
@@ -8,6 +8,7 @@ endif()
 
 map()
     kv(outofprocess true)
+    kv(locator lib${PLUGIN_WEBKITBROWSER_IMPLEMENTATION}.so)
     if(PLUGIN_WEBKITBROWSER_USER)
         kv(user ${PLUGIN_WEBKITBROWSER_USER})
     endif()
