+++ WebKitBrowser/WebKitBrowser.config
@@ -13,6 +13,7 @@ endif()
 
 map()
     kv(outofprocess true)
+    kv(locator lib${PLUGIN_WEBKITBROWSER_IMPLEMENTATION}.so)
     kv(mode "Container")
     if(PLUGIN_WEBKITBROWSER_USER)
         kv(user ${PLUGIN_WEBKITBROWSER_USER})
